code_interpreter/requirements.txt
redis==5.0.8
asyncpg==0.29.0
loguru==0.7.2
python_dotenv==1.0.1
sqlalchemy==2.0.32
